+++ src/extension.ts
@@ -25,10 +25,9 @@ export function activate(context: vscode.ExtensionContext) {
 		initTreeView();
 	}
 
-	if(vscode.workspace.name) {
-		vscode.commands.executeCommand('setContext', 'workspaceOpen', true);
-		
-	}
+	// if(vscode.workspace.name) {
+	// 	vscode.commands.executeCommand('setContext', 'workspaceOpen', true);		
+	// }
 
 	// Commands -------------------------------------------------------------------------
 
@@ -52,6 +51,22 @@ export function activate(context: vscode.ExtensionContext) {
 
 	context.subscriptions.push(vscode.commands.registerCommand('todoist.openCustomTask', (filePath : vscode.Uri, line : number, column : number) => {
 		openCustomTask(filePath, line, column);
+	}));
+
+	context.subscriptions.push(vscode.commands.registerCommand('todoist.addProject', () => {
+		vscode.window.showInputBox({
+			prompt: 'New Project Name',
+		}).then(projectName =>  {
+			if(projectName) {
+				vscode.window.showInformationMessage('Project : ' + projectName + ' is being created.');
+				const state = context.globalState;
+				const apiHelper = new todoistAPIHelper(state);
+				apiHelper.createProject(projectName).then(project => {
+					syncTodoist();
+					vscode.window.showInformationMessage('Project : ' + project.name + ' has been created.');
+				});
+			}
+		});
 	}));
 
 	// Event Handlers  -------------------------------------------------------------------